requirements.txt
cryptography
fastapi
gunicorn
opencv-python
torch
torchvision
numpy
onnx
onnxsim
onnxruntime
pandas
passlib[bcrypt]
python-decouple
python-jose
psycopg2-binary
sqlalchemy
tqdm
tensorboardX
scikit-learn
ultralytics
uvicorn